lasa_vs/requirements.txt
Flask==2.0.1
SpeechRecognition==3.8.1
pydub==0.25.1
nltk==3.6.3
python-Levenshtein==0.12.2
